+++ src/main/java/com/example/searchengine_ver1/core/widget/Widget.java
@@ -1,0 +1,6 @@
+package com.example.searchengine_ver1.core.widget;
+
+public interface Widget {
+    String getName();
+    String getImageUrl();
+}